fix(40088): extract navigation logic into useBridgeTransactionNavigation hook
